refactor(mirror): extract the tubafrenzy HTTP mirror client to @wxyc/legacy-mirror (BS#1707)
Moves the tubafrenzy HTTP mirror client (`http.mirror.ts`) and the active-rotation-match probe (`rotation-match.mirror.ts`) out of `apps/backend/middleware/legacy/` into a new shared `@wxyc/legacy-mirror` workspace, so the forthcoming `jobs/legacy-mirror-reconcile` cron (BS#1707) can re-drive orphaned show/entry rows through byte-identical payload mapping rather than a drift-prone re-implementation.

This is the enabling refactor (PR 1 of 2); it ships no behavior change.

- New `@wxyc/legacy-mirror` package holds the moved sources (via `git mv`, history preserved). It declares `@sentry/node` + `@wxyc/database` deps and a `drizzle-orm` peer — deliberately kept out of `@wxyc/database` itself, since this is an outbound REST + Sentry client, not ORM/schema access.
- `apps/backend/middleware/legacy/{http.mirror,rotation-match.mirror}.ts` are now thin re-export shims. `flowsheet.mirror.ts` and every other consumer keep their existing `./http.mirror.js` / `./rotation-match.mirror.js` import sites, so the collaborator unit tests that `jest.mock(...)` those paths keep intercepting unchanged.
- Build wiring: `@wxyc/legacy-mirror` added to the root `lint:prebuild` + `predev` build order (so its dist/types resolve before `apps/**` typecheck, matching how `@wxyc/lml-client` is wired) and to the `jest.unit.config.ts` source mapper.
- The two implementation test suites (`http.mirror.test.ts`, `rotation-match.mirror.test.ts`) are repointed to the moved module and the new sibling mock boundary.

Verified: full unit suite (4724 tests), typecheck across all workspaces, lint (0 errors), and the tubafrenzy mirror integration parity test.

Part of #1707.
