+++ Dockerfile.backend
@@ -25,6 +25,7 @@ COPY ./shared/database/package* ./shared/database/
 COPY ./shared/authentication/package* ./shared/authentication/
 COPY ./shared/lml-client/package* ./shared/lml-client/
 COPY ./shared/metadata/package* ./shared/metadata/
+COPY ./shared/legacy-mirror/package* ./shared/legacy-mirror/
 
 RUN npm install --omit=dev && rm -f .npmrc
 
@@ -33,6 +34,7 @@ COPY --from=builder ./builder/shared/database/dist ./shared/database/dist
 COPY --from=builder ./builder/shared/authentication/dist ./shared/authentication/dist
 COPY --from=builder ./builder/shared/lml-client/dist ./shared/lml-client/dist
 COPY --from=builder ./builder/shared/metadata/dist ./shared/metadata/dist
+COPY --from=builder ./builder/shared/legacy-mirror/dist ./shared/legacy-mirror/dist
 
 # Inherit the @wxyc/database 5s default. Backend HTTP handlers always finish
 # in milliseconds; anything longer than 5s is an orphan from a request that
