+++ package.json
@@ -8,7 +8,7 @@
   },
   "scripts": {
     "typecheck": "npm run lint:prebuild && npm run typecheck --workspace=@wxyc/database --workspace=shared/** --workspace=apps/**",
-    "lint:prebuild": "npm run build --workspace=@wxyc/database --workspace=@wxyc/authentication --workspace=@wxyc/lml-client --workspace=@wxyc/metadata",
+    "lint:prebuild": "npm run build --workspace=@wxyc/database --workspace=@wxyc/authentication --workspace=@wxyc/lml-client --workspace=@wxyc/metadata --workspace=@wxyc/legacy-mirror",
     "//lint": "8192 heap: eslint holds typed AST + program graph for every workspace; OOMs at node's default ~4GB ceiling on this monorepo.",
     "lint": "npm run lint:prebuild && NODE_OPTIONS=--max-old-space-size=8192 eslint --cache --cache-strategy content --cache-location .cache/eslint/ .",
     "lint:fix": "npm run lint:prebuild && NODE_OPTIONS=--max-old-space-size=8192 eslint --cache --cache-strategy content --cache-location .cache/eslint/ . --fix",
@@ -16,7 +16,7 @@
     "format": "prettier --cache --cache-location .cache/prettier/ --write .",
     "format:check": "prettier --cache --cache-location .cache/prettier/ --check .",
     "build": "npm run build --workspace=@wxyc/database --workspace=shared/** --workspace=apps/** --workspace=jobs/**",
-    "predev": "npm run build --workspace=@wxyc/database --workspace=@wxyc/authentication --workspace=@wxyc/lml-client --workspace=@wxyc/metadata",
+    "predev": "npm run build --workspace=@wxyc/database --workspace=@wxyc/authentication --workspace=@wxyc/lml-client --workspace=@wxyc/metadata --workspace=@wxyc/legacy-mirror",
     "dev": "dotenvx run -f .env -- concurrently \"npm:dev:auth\" \"npm:dev:backend\"",
     "dev:backend": "npm run dev --workspace=@wxyc/backend",
     "dev:auth": "npm run dev --workspace=@wxyc/auth-service",
