+++ shared/legacy-mirror/tsconfig.build.json
@@ -1,0 +1,7 @@
+{
+  "extends": ["./tsconfig.json"],
+  "compilerOptions": {
+    "composite": false,
+    "incremental": false
+  }
+}
